requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asn1crypto==1.5.1
babel==2.16.0
bcrypt==4.2.0
bleach==6.1.0
blinker==1.8.2
cbor2==5.6.5
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
dnspython==2.7.0
email_validator==2.2.0
Flask==3.0.3
flask-babel==4.0.0
Flask-Login==0.6.3
Flask-Mailman==1.1.1
Flask-Principal==0.4.0
Flask-Security==5.5.2
flask-swagger-ui==4.11.1
Flask-WTF==1.2.2
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.1
passlib==1.7.4
phonenumberslite==8.13.48
pillow==11.0.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pytest==8.3.3
python-dotenv==1.0.1
pytz==2024.2
qrcode==8.0
redis==5.1.1
setuptools==75.2.0
six==1.16.0
webauthn==2.2.0
webencodings==0.5.1
Werkzeug==3.0.4
WTForms==3.2.1
